+++ src/global.ts
@@ -6,6 +6,26 @@ import type {
 import type { EventTimer } from '.';
 
 declare global {
+	type ConnectionType =
+		| 'bluetooth'
+		| 'cellular'
+		| 'ethernet'
+		| 'mixed'
+		| 'none'
+		| 'other'
+		| 'unknown'
+		| 'wifi';
+
+	interface NetworkInformation extends EventTarget {
+		readonly type?: ConnectionType;
+		readonly downlink?: number;
+		readonly effectiveType?: string;
+	}
+
+	interface Navigator {
+		readonly connection: NetworkInformation;
+	}
+
 	interface Window {
 		google_trackConversion?: (arg0: GoogleTrackConversionObject) => void;
 		google_tag_params?: GoogleTagParams;
@@ -18,5 +38,6 @@ declare global {
 			config: GuardianWindowConfig;
 		};
 		ga: UniversalAnalytics.ga | null;
+		readonly navigator: Navigator;
 	}
 }
